fix: duplicate

+++ src/screens/login/index.tsx
@@ -20,10 +20,17 @@ import DBCollections from '../../services/sqlite/collections'
 import { IndexProps } from "./models";
 import View from "./view";
 import { sincronizeDB } from "../../constants/sincronize";
+import { CollectionsTypes } from "../../services/redux/reducers/collections/models";
+import { ItemsType } from "../../services/api/collections/models";
+import { SuppliersTypes } from "../../services/redux/reducers/suppliers/models";
+import { ProductsProps } from "../../services/sqlite/products/models";
+import NetInfo from "@react-native-community/netinfo";
 
 const Login: React.FC<IndexProps> = ({
     userData,
-    setLoginData
+    setLoginData,
+    isSincronized,
+    setIsSincronized
 }) => {
     const navigation = useNavigation<NativeStackNavigationProp<StackProps>>()
 
@@ -45,6 +52,7 @@ const Login: React.FC<IndexProps> = ({
     const [showPassword, setShowPassword] = useState<boolean>(false)
 
     const [isLoading, setIsLoading] = useState<boolean>(false)
+    const [isLoadingSincronized, setIsLoadingSincronized] = useState<boolean>(false)
 
     useEffect(() => {
         loadData()
@@ -57,8 +65,84 @@ const Login: React.FC<IndexProps> = ({
         }
 
         users.listStores().then((data: StoresProps[]) => setStores(data))
-        DBCollections.deleteAll()
-        DBProducts.deleteAll()
+    }
+
+    const start = async () => {
+        NetInfo.fetch().then(state => {
+            if (!state.isConnected) {
+                Alert.alert('Aviso', 'Seu dispositivo está sem internet, por favor, conecte-o para acessar o aplicativo')
+                return
+            }
+        });
+
+        if (!isSincronized) {
+            setIsLoadingSincronized(true)
+            await DBCollections.listSuccess()
+            .then(async (data: SuppliersTypes['data'][]) => {
+                data.map(item => {
+                    let codOrdemColeta: string, DTULAlteracao: string, dtHoraColeta: string, data: string, dtHoraStatus: string, numCar: string
+                    let qtTotalColetada: string, qtItensColetados: string, qtItensPrevistos: string, qtPrevista: string, pesoColeta: string
+                    let vlTotal: string, vlColeta: string, newItems: ItemsType[]
+                    
+                    Promise.all([
+                        DBProducts.listAll({ CODFORNEC: item.CODFORNEC })
+                        .then((dataP: ProductsProps[]) => {
+                            codOrdemColeta = item.CODORDEMCOLETA
+                            DTULAlteracao = item.DTULTALTERACAO
+                            dtHoraColeta = item.DTCOLETA
+                            data = item.DTCOLETA
+                            dtHoraStatus = item.DTCOLETA
+                            numCar = userData.numCar
+
+                            qtTotalColetada = String(dataP.reduce((accumulattor, total) => Number(accumulattor) + Number(total.COLETA), 0))
+                            qtItensColetados = String(dataP.reduce((accumulattor, total) => Number(total) > 0 ? Number(accumulattor) + 1 : 1, 0))
+                            qtItensPrevistos = String(dataP.length)
+                            qtPrevista = String(dataP.reduce((valor, total) => Number(valor) + Number(total.QTPREVISAO), 0))
+                            pesoColeta = String(dataP.reduce((accumulattor, total) => Number(accumulattor) + Number(total.COLETA), 0))
+
+                            vlTotal = String(dataP.reduce((accumulattor, total) => Number(accumulattor) + (Number(total.PCOMPRA || 0) * Number(total.COLETA || 0)), 0))
+                            vlColeta = String(dataP.reduce((accumulattor, total) => Number(accumulattor) + (Number(total.PCOMPRA || 0) * Number(total.COLETA || 0)), 0))
+
+                            newItems = dataP.map(item => ({
+                                codProd: String(item.CODPROD),
+                                quantidade: item.COLETA,
+                                pCompra: item.PCOMPRA,
+                                qtPrevista: item.QTPREVISAO,
+                            }))
+                    
+                        })
+                    ]).then(() => {
+                        collections.addItem({
+                            itens: newItems,
+                            codOrdemColeta,
+                            DTULAlteracao,
+                            qtTotalColetada,
+                            qtItensColetados,
+                            qtItensPrevistos,
+                            vlTotal,
+                            qtPrevista,
+                            dtHoraColeta,
+                            data,
+                            dtHoraStatus,
+                            pesoColeta,
+                            vlColeta,
+                            numCar
+                        })
+                    }).then(() => {
+                        DBCollections.deleteAll()
+                        DBProducts.deleteAll()
+                    })
+                })
+            })
+        } else {
+            DBCollections.deleteAll()
+            DBProducts.deleteAll()
+        }
+
+        setIsSincronized(true)
+        setIsLoadingSincronized(false)
+
+        setShowModal(true)
     }
 
     useEffect(() => {
@@ -126,6 +210,7 @@ const Login: React.FC<IndexProps> = ({
 
     return (
         <View
+            start={start}
             showModal={showModal}
             setShowModal={setShowModal}
             showStores={showStores}
@@ -149,20 +234,25 @@ const Login: React.FC<IndexProps> = ({
             showPassword={showPassword}
             setShowPassword={setShowPassword}
             isLoading={isLoading}
+            isLoadingSincronized={isLoadingSincronized}
         />
     )
 }
 
 const mapStateToProps = ({
-    loginReducer
+    loginReducer,
+    collectionsReducer,
 }: {
-    loginReducer: LoginTypes
+    loginReducer: LoginTypes,
+    collectionsReducer: CollectionsTypes,
 }) => ({
     userData: loginReducer.data,
+    isSincronized: collectionsReducer.sincronized
 })
 
 const mapDispatchToProps = (dispatch: Dispatch) => ({
     setLoginData: (data: LoginTypes['data']) => dispatch({ type: 'SET_LOGIN_DATA', payload: { data } }),
+    setIsSincronized: (data: CollectionsTypes['sincronized']) => dispatch({ type: 'SET_SINCRONIZED', payload: { data } }),
 })
 
 export default connect(mapStateToProps, mapDispatchToProps)(Login);
